Video 156 - Attackers Stricking Target

+++ Assets/Scripts/Health.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 
 public class Health : MonoBehaviour
 {
-    [SerializeField] int health = 100;
+    [SerializeField] float health = 100;
     [SerializeField] GameObject deathVFX;
 
-    public void DealDamage(int damage)
+    public void DealDamage(float damage)
     {
         health -= damage;
         if (health <= 0)
